# docs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "EventHouse.Management.Api",
    "version": "v1"
  },
  "servers": [
    {
      "url": "https://eventhouse-management-api-demo.onrender.com",
      "description": "Render Production"
    },
    {
      "url": "http://localhost:5185",
      "description": "Local"
    }
  ],
  "paths": {
...
    "/api/v1/events/{eventId}": {
      "get": {
        "tags": [
          "Events"
        ],
        "summary": "Retrieve a specific event by its unique identifier.",
        "operationId": "GetEventById",
        "parameters": [
          {
            "name": "eventId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventResponse"
                },
                "example": {
                  "id": "b123c456-8d89-1e12-4f45-7a7890123457",
                  "name": "Summer Fest 2026",
                  "description": "Annual open-air music festival.",
                  "scope": "International"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        }
      },
      "put": {
        "tags": [
          "Events"
        ],
        "summary": "Update an existing event in the system.",
        "operationId": "UpdateEvent",
        "parameters": [
          {
            "name": "eventId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-Idempotency-Key",
            "in": "header",
            "description": "Idempotency header configured. Allowed methods: POST, PUT, DELETE",
            "schema": {
              "type": "string",
              "format": "uuid",
              "default": "13996d86-8ff2-4a8b-b9e5-1def7a86ec9d"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEventRequest"
              },
              "example": {
                "name": "Summer Fest 2026",
                "description": "Annual open-air music festival.",
                "scope": "International"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEventRequest"
              },
              "example": {
                "name": "Summer Fest 2026",
                "description": "Annual open-air music festival.",
                "scope": "International"
              }
            },
            "application/*\u002Bjson": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEventRequest"
              },
              "example": {
                "name": "Summer Fest 2026",
                "description": "Annual open-air music festival.",
                "scope": "International"
              }
            }
          }
        },
        "responses": {
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "204": {
            "description": "No Content"
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        }
      },
      "delete": {
        "tags": [
          "Events"
        ],
        "summary": "Delete an existing event from the system.",
        "operationId": "DeleteEvent",
        "parameters": [
          {
            "name": "eventId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-Idempotency-Key",
            "in": "header",
            "description": "Idempotency header configured. Allowed methods: POST, PUT, DELETE",
            "schema": {
              "type": "string",
              "format": "uuid",
              "default": "1614922d-22dc-435d-a6bb-434acf067a51"
            }
          }
        ],
        "responses": {
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "204": {
            "description": "No Content"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        }
      }
    },
...
    "/api/v1/event-venues/{eventVenueId}": {
      "get": {
        "tags": [
          "EventVenues"
        ],
        "summary": "Retrieve a specific event venue by their unique identifier.",
        "operationId": "GetEventVenueById",
        "parameters": [
          {
            "name": "eventVenueId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventVenueResponse"
                },
                "example": {
                  "id": "a123b456-7c89-0d12-3e45-6f7890123456",
                  "eventId": "b123c456-8d89-1e12-4f45-7a7890123457",
                  "venueId": "c123d456-9e89-2f12-5a45-8b7890123458",
                  "status": "Active",
                  "eventName": "Summer Fest 2026",
                  "venueName": "Madison Square Garden"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        }
      }
    },
    "/api/v1/event-venues": {
      "get": {
        "tags": [
          "EventVenues"
        ],
        "summary": "List event venues with optional filtering, sorting, and pagination.",
        "operationId": "ListEventVenues",
        "parameters": [
          {
            "name": "eventId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "venueId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/EventVenueStatus"
            }
          },
          {
            "name": "sortBy",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/EventVenueSortBy"
            }
          },
          {
            "name": "sortDirection",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/SortDirection"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "maximum": 2147483647,
              "minimum": 1,
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "maximum": 100,
              "minimum": 1,
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventVenueResponsePagedResult"
                },
                "example": {
                  "items": [
                    {
                      "id": "a123b456-7c89-0d12-3e45-6f7890123456",
                      "eventId": "b123c456-8d89-1e12-4f45-7a7890123457",
                      "venueId": "c123d456-9e89-2f12-5a45-8b7890123458",
                      "status": "Active",
                      "eventName": "Summer Fest 2026",
                      "venueName": "Madison Square Garden"
                    }
                  ],
                  "totalCount": 1,
                  "page": 1,
                  "pageSize": 20,
                  "links": {
                    "self": "/api/v1/event-venues?page=1\u0026pageSize=20",
                    "first": "/api/v1/event-venues?page=1\u0026pageSize=20",
                    "previous": null,
                    "next": null,
                    "last": "/api/v1/event-venues?page=1\u0026pageSize=20"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        }
      },
      "post": {
        "tags": [
          "EventVenues"
        ],
        "summary": "Create a new event venue in the system.",
        "operationId": "CreateEventVenue",
        "parameters": [
          {
            "name": "X-Idempotency-Key",
            "in": "header",
            "description": "Idempotency header configured. Allowed methods: POST, PUT, DELETE",
            "schema": {
              "type": "string",
              "format": "uuid",
              "default": "b49ca17c-8192-438c-b384-2589e95572a8"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEventVenueRequest"
              },
              "example": {
                "eventId": "b123c456-8d89-1e12-4f45-7a7890123457",
                "venueId": "c123d456-9e89-2f12-5a45-8b7890123458",
                "status": "Active"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEventVenueRequest"
              },
              "example": {
                "eventId": "b123c456-8d89-1e12-4f45-7a7890123457",
                "venueId": "c123d456-9e89-2f12-5a45-8b7890123458",
                "status": "Active"
              }
            },
            "application/*\u002Bjson": {
              "schema": {
                "$ref": "#/components/schemas/CreateEventVenueRequest"
              },
              "example": {
                "eventId": "b123c456-8d89-1e12-4f45-7a7890123457",
                "venueId": "c123d456-9e89-2f12-5a45-8b7890123458",
                "status": "Active"
              }
            }
          }
        },
        "responses": {
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "201": {
            "description": "Created",
            "headers": {
              "Location": {
                "description": "URL of the newly created resource",
                "schema": {
                  "type": "string",
                  "format": "uri"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventVenueResponse"
                },
                "example": {
                  "id": "a123b456-7c89-0d12-3e45-6f7890123456",
                  "eventId": "b123c456-8d89-1e12-4f45-7a7890123457",
                  "venueId": "c123d456-9e89-2f12-5a45-8b7890123458",
                  "status": "Active",
                  "eventName": "Summer Fest 2026",
                  "venueName": "Madison Square Garden"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        }
      }
    },
    "/api/v1/event-venues/{eventVenueId}/status": {
      "put": {
        "tags": [
          "EventVenues"
        ],
        "summary": "Update an existing event venues details.",
        "operationId": "UpdateEventVenueStatus",
        "parameters": [
          {
            "name": "eventVenueId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-Idempotency-Key",
            "in": "header",
            "description": "Idempotency header configured. Allowed methods: POST, PUT, DELETE",
            "schema": {
              "type": "string",
              "format": "uuid",
              "default": "11e3620e-f116-4719-a717-4ff738571121"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEventVenueStatusRequest"
              },
              "example": {
                "status": "Active"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEventVenueStatusRequest"
              },
              "example": {
                "status": "Active"
              }
            },
            "application/*\u002Bjson": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEventVenueStatusRequest"
              },
              "example": {
                "status": "Active"
              }
            }
          }
        },
        "responses": {
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "204": {
            "description": "No Content"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        }
      }
    },
...
    "/api/v1/venues/{venueId}": {
      "get": {
        "tags": [
          "Venues"
        ],
        "summary": "Retrieve a specific venue by their unique identifier.",
        "operationId": "GetVenueById",
        "parameters": [
          {
            "name": "venueId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VenueResponse"
                },
                "example": {
                  "id": "c123d456-9e89-2f12-5a45-8b7890123458",
                  "name": "Madison Square Garden",
                  "address": "4 Pennsylvania Plaza, New York, NY 10001",
                  "city": "New York",
                  "region": "NY",
                  "countryCode": "US",
                  "latitude": 40.7505,
                  "longitude": -73.9934,
                  "timeZoneId": "America/New_York",
                  "capacity": 20000,
                  "isActive": true
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        }
      },
      "put": {
        "tags": [
          "Venues"
        ],
        "summary": "Update an existing venue\u0027s details.",
        "operationId": "UpdateVenue",
        "parameters": [
          {
            "name": "venueId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-Idempotency-Key",
            "in": "header",
            "description": "Idempotency header configured. Allowed methods: POST, PUT, DELETE",
            "schema": {
              "type": "string",
              "format": "uuid",
              "default": "7eb2f1b0-7279-4373-9dc8-49e94da53abb"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateVenueRequest"
              },
              "example": {
                "name": "Madison Square Garden",
                "address": "4 Pennsylvania Plaza, New York, NY 10001",
                "city": "New York",
                "region": "NY",
                "countryCode": "US",
                "latitude": 40.7505,
                "longitude": -73.9934,
                "timeZoneId": "America/New_York",
                "capacity": 20000,
                "isActive": true
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateVenueRequest"
              },
              "example": {
                "name": "Madison Square Garden",
                "address": "4 Pennsylvania Plaza, New York, NY 10001",
                "city": "New York",
                "region": "NY",
                "countryCode": "US",
                "latitude": 40.7505,
                "longitude": -73.9934,
                "timeZoneId": "America/New_York",
                "capacity": 20000,
                "isActive": true
              }
            },
            "application/*\u002Bjson": {
              "schema": {
                "$ref": "#/components/schemas/UpdateVenueRequest"
              },
              "example": {
                "name": "Madison Square Garden",
                "address": "4 Pennsylvania Plaza, New York, NY 10001",
                "city": "New York",
                "region": "NY",
                "countryCode": "US",
                "latitude": 40.7505,
                "longitude": -73.9934,
                "timeZoneId": "America/New_York",
                "capacity": 20000,
                "isActive": true
              }
            }
          }
        },
        "responses": {
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "204": {
            "description": "No Content"
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        }
      },
      "delete": {
        "tags": [
          "Venues"
        ],
        "summary": "Delete a venue from the system.",
        "operationId": "DeleteVenue",
        "parameters": [
          {
            "name": "venueId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-Idempotency-Key",
            "in": "header",
            "description": "Idempotency header configured. Allowed methods: POST, PUT, DELETE",
            "schema": {
              "type": "string",
              "format": "uuid",
              "default": "910f3adc-a6fd-40b8-b5fc-15af60228238"
            }
          }
        ],
        "responses": {
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "204": {
            "description": "No Content"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "EventResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier of the event.",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "description": "Event name."
          },
          "description": {
            "type": "string",
            "description": "Optional event description.",
            "nullable": true
          },
          "scope": {
            "$ref": "#/components/schemas/EventScope"
          }
        },
        "additionalProperties": false
      },
...
      "EventVenueResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "The unique identifier of the event venue.",
            "format": "uuid"
          },
          "eventId": {
            "type": "string",
            "description": "The unique identifier of the event to be associated with the event venue.",
            "format": "uuid"
          },
          "venueId": {
            "type": "string",
            "description": "The unique identifier of the venue to be associated with the event venue.",
            "format": "uuid"
          },
          "status": {
            "$ref": "#/components/schemas/EventVenueStatus"
          },
          "eventName": {
            "type": "string",
            "description": "Gets the name of the event.",
            "nullable": true
          },
          "venueName": {
            "type": "string",
            "description": "Gets the name of the venue.",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "EventVenueResponsePagedResult": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EventVenueResponse"
            }
          },
          "totalCount": {
            "type": "integer",
            "format": "int32"
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "links": {
            "$ref": "#/components/schemas/PaginationLinks"
          }
        },
        "additionalProperties": false
      },
      "EventVenueSortBy": {
        "enum": [
          "Status"
        ],
        "type": "string"
      },
      "EventVenueStatus": {
        "enum": [
          "Active",
          "Inactive"
        ],
        "type": "string"
      },
...
      "SortDirection": {
        "enum": [
          "Asc",
          "Desc"
        ],
        "type": "string"
      },
...
      "UpdateEventRequest": {
        "required": [
          "name",
          "scope"
        ],
        "type": "object",
        "properties": {
          "name": {
            "maxLength": 200,
            "minLength": 2,
            "type": "string",
            "description": "Event name."
          },
          "description": {
            "type": "string",
            "description": "Optional event description.",
            "nullable": true
          },
          "scope": {
            "$ref": "#/components/schemas/EventScope"
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "description": "JWT Authorization header. Example: Bearer {token}",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Unauthorized",
        "content": {
          "application/problem\u002Bjson": {
            "schema": {
              "$ref": "#/components/schemas/EventHouseProblemDetails"
            },
            "example": {
              "type": "urn:eventhouse:error:UNAUTHORIZED",
              "title": "Unauthorized",
              "status": 401,
              "detail": "A valid JWT access token is required.",
              "errorCode": "UNAUTHORIZED",
              "traceId": "00-aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa-bbbbbbbbbbbbbbbb-01"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Forbidden",
        "content": {
          "application/problem\u002Bjson": {
            "schema": {
              "$ref": "#/components/schemas/EventHouseProblemDetails"
            },
            "example": {
              "type": "urn:eventhouse:error:FORBIDDEN",
              "title": "Forbidden",
              "status": 403,
              "detail": "You do not have permission to access this resource.",
              "errorCode": "FORBIDDEN",
              "traceId": "00-aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa-bbbbbbbbbbbbbbbb-01"
            }
          }
        }
      },
      "InternalError": {
        "description": "Internal Server Error",
        "content": {
          "application/problem\u002Bjson": {
            "schema": {
              "$ref": "#/components/schemas/EventHouseProblemDetails"
            },
            "example": {
              "type": "urn:eventhouse:error:INTERNAL_ERROR",
              "title": "Internal Server Error",
              "status": 500,
              "detail": "An unexpected error occurred.",
              "errorCode": "INTERNAL_ERROR",
              "traceId": "00-aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa-bbbbbbbbbbbbbbbb-01"
            }
          }
        }
      },
      "NotFound": {
        "description": "Not Found",
        "content": {
          "application/problem\u002Bjson": {
            "schema": {
              "$ref": "#/components/schemas/EventHouseProblemDetails"
            },
            "example": {
              "type": "urn:eventhouse:error:NOT_FOUND",
              "title": "Not Found",
              "status": 404,
              "detail": "The requested resource was not found.",
              "errorCode": "NOT_FOUND",
              "traceId": "00-aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa-bbbbbbbbbbbbbbbb-01"
            }
          }
        }
      },
      "Conflict": {
        "description": "Conflict",
        "content": {
          "application/problem\u002Bjson": {
            "schema": {
              "$ref": "#/components/schemas/EventHouseProblemDetails"
            },
            "example": {
              "type": "urn:eventhouse:error:CONFLICT",
              "title": "Conflict",
              "status": 409,
              "detail": "The request conflicts with the current state of the resource.",
              "errorCode": "CONFLICT",
              "traceId": "00-aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa-bbbbbbbbbbbbbbbb-01"
            }
          }
        }
      },
      "ValidationError": {
        "description": "Validation error",
        "content": {
          "application/problem\u002Bjson": {
            "schema": {
              "$ref": "#/components/schemas/ValidationProblemDetails"
            },
            "example": {
              "type": "urn:eventhouse:error:VALIDATION_ERROR",
              "title": "Validation error",
              "status": 400,
              "detail": "One or more validation errors occurred.",
              "errors": {
                "name": [
                  "The Name field is required."
                ]
              },
              "traceId": "00-aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa-bbbbbbbbbbbbbbbb-01"
            }
          }
        }
      },
      "TooManyRequests": {
        "description": "Too Many Requests",
        "headers": {
          "Retry-After": {
            "description": "Seconds to wait before retrying the request.",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        },
        "content": {
          "application/problem\u002Bjson": {
            "schema": {
              "$ref": "#/components/schemas/EventHouseProblemDetails"
            },
            "example": {
              "type": "urn:eventhouse:error:RATE_LIMIT_EXCEEDED",
              "title": "Too Many Requests",
              "status": 429,
              "detail": "Rate limit exceeded. Please retry later.",
              "errorCode": "RATE_LIMIT_EXCEEDED",
              "traceId": "00-aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa-bbbbbbbbbbbbbbbb-01"
            }
          }
        }
      }
    }
  },
  "security": [
    {
      "bearer": []
    }
  ]
}